@@ -1,7 +1,12 @@
-﻿using Microsoft.AspNetCore;
+using System;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.DependencyInjection;
-using Notification.Service;
+using Microsoft.Extensions.Hosting;
+using Notification.Interfaces;
+using Notification.Services;
+using Serilog;
+using Serilog.Events;
 
 namespace Notification
 {
@@ -9,14 +14,35 @@ namespace Notification
     {
         public static void Main(string[] args)
         {
-            var host = CreateWebHostBuilder(args).Build();
-            host.Services.GetRequiredService<INotificationService>().Start();
-            host.Run();
-        }
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Information()
+                .MinimumLevel.Override("Microsoft", LogEventLevel.Warning)
+                .Enrich.FromLogContext()
+                .CreateLogger();
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
+            try
+            {
+                var host = CreateHostBuilder(args).Build();
+                host.Services.GetRequiredService<INotificationService>().Start();
+                host.Run();
+            }
+            catch (Exception ex)
+            {
+                Log.Fatal(ex, "Host terminated unexpectedly");
+            }
+            finally
+            {
+                Log.CloseAndFlush();
+            }
+        }
+        public static IHostBuilder CreateHostBuilder(string[] args)
+        {
+            return Host.CreateDefaultBuilder(args)
+                .ConfigureServices((context, services) => {
+                    services.Configure<KestrelServerOptions>(context.Configuration.GetSection("Kestrel"));
+                })
+                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); })
                 .UseDefaultServiceProvider(options => options.ValidateScopes = false);
+        }
     }
 }
